updates of R scripts, error checks

+++ addScripts/BactFunctions.py
@@ -113,31 +113,40 @@ def concatNuc(RVDfiles, RESULTSfiles):
 
 #Checks CSV files to be used for bound matrix for missing columns based on FASTAlist
 def csvFix(Rfiles, FASTAlist):
-    matCSV = pd.read_csv(Rfiles + "matTRF.csv", index_col=0)
-    orthoCSV = pd.read_csv(Rfiles + "OrthoGCMatrix.csv", index_col=0)
-    groupCSV = pd.read_csv(Rfiles + "distal_GroupMatrix.csv", index_col=0)
     FASTAprefix = []
     for name in FASTAlist:
         temp = name.split(".")
         FASTAprefix.append(temp[0])
-    matCols = list(matCSV.columns)
-    orthoCols = list(orthoCSV.columns)
-    groupCols = list(groupCSV.columns)
-    matCompare = (list(set(FASTAprefix).difference(matCols)))
-    orthoCompare = (list(set(FASTAprefix).difference(orthoCols)))
-    groupCompare = (list(set(FASTAprefix).difference(groupCols)))
-    for col in matCompare:
-        print("Missing " + col + " in matTRF.csv. Adding, setting values to 0")
-        matCSV[col] = int(0)
-    for col in orthoCompare:
-        print("Missing " + col + " in OrthoGCMatrix.csv. Adding, setting values to 0")
-        orthoCSV[col] = int(0)
-    for col in groupCompare:
-        print("Missing " + col + " in distal_GroupMatrix.csv. Adding, setting values to 0")
-        groupCSV[col] = int(0)
-    matCSV.to_csv(Rfiles + "matTRF.csv")
-    orthoCSV.to_csv(Rfiles + "OrthoGCMatrix.csv")
-    groupCSV.to_csv(Rfiles + "distal_GroupMatrix.csv")
+    if(os.path.isfile(Rfiles + "matTRF.csv")):
+        matCSV = pd.read_csv(Rfiles + "matTRF.csv", index_col=0)
+        matCols = list(matCSV.columns)
+        matCompare = (list(set(FASTAprefix).difference(matCols)))
+        for col in matCompare:
+            print("Missing " + col + " in matTRF.csv. Adding, setting values to 0")
+            matCSV[col] = int(0)
+        matCSV.to_csv(Rfiles + "matTRF.csv")
+    else:
+        print("matTRF.csv was not generated, not applying csvFix")
+    if(os.path.isfile(Rfiles + "OrthoGCMatrix.csv")):
+        orthoCSV = pd.read_csv(Rfiles + "OrthoGCMatrix.csv", index_col=0
+        orthoCols = list(orthoCSV.columns)
+        orthoCompare = (list(set(FASTAprefix).difference(orthoCols)))
+        for col in orthoCompare:
+            print("Missing " + col + " in OrthoGCMatrix.csv. Adding, setting values to 0")
+            orthoCSV[col] = int(0)
+        orthoCSV.to_csv(Rfiles + "OrthoGCMatrix.csv")
+    else:
+        print("OrthoGCMatrix.csv was not generated, not applying csvFix")
+    if(os.path.isfile(Rfiles + "distal_GroupMatrix.csv")):
+        groupCSV = pd.read_csv(Rfiles + "distal_GroupMatrix.csv", index_col=0)
+        groupCols = list(groupCSV.columns)
+        groupCompare = (list(set(FASTAprefix).difference(groupCols)))
+        for col in groupCompare:
+            print("Missing " + col + " in distal_GroupMatrix.csv. Adding, setting values to 0")
+            groupCSV[col] = int(0)
+        groupCSV.to_csv(Rfiles + "distal_GroupMatrix.csv")
+    else:
+        print("distal_GroupMatrix.csv was not generated, not applying csvFix")
 
 
 #Function for initializing scoary and second round of R modifications
@@ -145,17 +154,20 @@ def scoary(pipePath, providedCSV):
     shutil.copyfile(providedCSV, pipePath + "SCOARYfiles/providedCSV.csv")
     scorFile = (pipePath + "SCOARYfiles/providedCSV.csv")
     if scorFile.lower().endswith(".csv"):
-        boundCSV = pd.read_csv(pipePath + "Rfiles/boundMatrix.csv", index_col=0)
-        traitCSV = pd.read_csv(scorFile, index_col=0)
-        boundCols = list(boundCSV.columns)
-        traitRows = list(traitCSV.index)
-        boundTraitCompare = list(set(boundCols) - set(traitRows))
-        if 'V1' in boundTraitCompare:
-            boundTraitCompare.remove('V1')
-        if len(boundTraitCompare) is 0:
-            return scorFile;
+        if(os.path.exists(pipePath + "Rfiles/boundMatrix.csv")):
+            boundCSV = pd.read_csv(pipePath + "Rfiles/boundMatrix.csv", index_col=0)
+            traitCSV = pd.read_csv(scorFile, index_col=0)
+            boundCols = list(boundCSV.columns)
+            traitRows = list(traitCSV.index)
+            boundTraitCompare = list(set(boundCols) - set(traitRows))
+            if 'V1' in boundTraitCompare:
+                boundTraitCompare.remove('V1')
+            if len(boundTraitCompare) is 0:
+                return scorFile;
+            else:
+                print("Columns of boundMatrix do not match rows of provided CSV, unable to run Scoary")
         else:
-            print("Columns of boundMatrix do not match rows of provided CSV, unable to run Scoary")
+            print("Missing boundMatrix, unable to run Scoary")
     else:
         print("Provided csv does not end with csv extension, unable to proceed.")
